Implemented new semantics for ptionalness & pages:
The 'optional' attribute of a page is only used to inherit from by contained fields (see newField()), at runtime it doesn't have meaning in itself because whether or not a page can be skipped or left is to be decided based on the optionalness and acquired values of the contained field. Because of this the optionalness of the page is reset to ALWAYS after all contained fields are parsed.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>