Replaced DateTimeColumn by new TimeStampColumn: - the new TimeStamp class holds the milliseconds, the UTC offset (as quarter hours) and transient(!) DateTime object - the advantage is we avoid storing the very complex DateTime objects in DB4O and they are only created in runtime memory when needed. - this fixes the problem where DateTime objects retrieved from DB4O would not be equal() to freshly instantiate ones with the exact same time and timezone (likely due to bug/shortcoming in JodaTime). This in turn caused EqualityConstraints to fail on such pairs of DateTime object. And that in turn meant we couldn't use EqualityConstraints to construct the query returned by ForeignKey#getForeignRecordQuery() - instead we used RuleConstraints but that was risky because those only work on ComparatorColumns and there is no guarantee indexes will only contain such columns.
Note 1: the above-mentioned "shortcoming" in JodaTime is discussed here: https://github.com/JodaOrg/joda-time/issues/36 (reported by me). The issue has supposedly been fixed in v2.3 (see replies) but that turns out to be incorrect (or at least the fix they did was incomplete), when I find some timer I may investigate this further to send them a follow-up bugreport.

Note 2: For now I'm committing this as a separate branch. The reason is this is only really needed when using DB4O in combination with (the current version of) JodaTime. Once we ditch DB4O most of the benefits of the TimeStamp(Column) over the old DateTimeColumn will disappear. Nevertheless I will probably merge this into the master soon. We can always updo the changes later if we want (should not be very hard).

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>