First serious attempt at Mavenising the Collector & Library projects. Several issues remain however: - loading of library doesn't work in eclipse (loading the project also as a project dependency solves compile-time problems but the library remains missing when an APK is being built...) - environment2 & filechooser not yet loaded via maven-android-sdk-deployer
On the upside, exploring the possibilities of maven has led me to finally solve the annoying problem in which he BuildInfo.java file was not regenerated after a project clean (due to a known bug in Eclipse: Ant builders not running after clean).
Now the BuildInfo is generate by a project builder which runs mvn which in turn runs the ant script. This only works if a stand-alone Maven installation is present on the system PATH though. Moreover because Eclipse isn't as smart as cmd.exe we need to execute "mvn.bat" (instead of just "mvn", which is resolved to "mvn.bat" when typed into cmd.exe), annoyingly this causes the Collector project configuration to become platform-dependent. However, on OSX or Linux it should suffice to change "mvn.bat" into "mvn" in the builder config to make it work.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>