Improvements to parsing exception handling to deal with stupidity of SAXException class:
The class can take another Exception upon construction but doesn't treat it like a "cause" in the way the Exception base class expects. As a result causes are not printed as part of stacktraces. To solve this DocumentParser#parse() now unpacks SAXExceptions and turns them into regular Exception with a proper cause before throwing them. I've also removed usages of SAXException with a cause exception and message in FormParser & ProjectParser, replacing them by regular Exceptions.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>