Changes to the way we deal with the AndroidManifest.xml:  - the "source" manifest (in the project root) no longer has the versionName and versionCode which will appear in the app, instead the are set by maven/AMP;  - however maven/AMP no longer modify the "source" manifest  - in eclipse AMP will update the manifest in bin/  - command-line builds will reproduce it in target/.
All builds use the project version (as defined in pom.xml's!) as the versionName, but eclipse builds get an "-eclipse" suffix (may remove this later though).
The versionCode is a more complicated matter... If we use AMPs <versionCodeUpdateFromVersion> feature all beta's/rc's of a given version will get the same versionCode (e.g. v2.0.0-beta-2 and v2.0.0-beta25 both get 2000000). This is fine for testing but not for releases on the Play store which differentiates version by versionCode! To work around this we only use <versionCodeUpdateFromVersion> on eclipse builds. Command-line builds use a time-based versionCode value instead (= the number of hours since 2012-01-01T00:00:00.000Z. This means we can theoretically release a new build to the Play store every hour ;-).

Also updated to AMP v4.1.1.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>