ProjectManagerActivity no longer restarts upon screen orientation/size and keyboard changes. This effectively resolves issue #21 and likely has other benefits (no reloading of the project list).
I should note though that the underlying problem of issue #21 is that the AsyncDownloader is unaware of activities being restarted while it is running, which means it tries to dismiss a progess dialog which is not longer on the screen (and owned by a finished/destroyed activity). I have not resolved this root cause. I did add notes to javadoc to warn about it and added basic exception handling to avoid actual crashes. Still, trouble may likely ensure if AsyncDownloader were to be used in combination with an activity which restarted upon screen orientation changes (for example). A *proper* solution would probably involve using an AsyncTaskLoader instead of an AsyncTask (see: http://stackoverflow.com/questions/18214293).

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>