Some work on separate BroadcastReceivers for SMS sent & delivered event. (Incomplete)
These will have to query the Tstore to get the Transmission/Message object (to get its callback). To make this efficient in the general case where sending happens almost immediately we'll need to give the Tstore a caching mechanism such that we can return the same Transmission/Message instance which was passed to AndroidSMSClient when the SMS was being sent.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>